% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{Alpha}
\alias{Alpha}
\title{The Alpha function is to generate valid significant level grid values}
\usage{
Alpha(r, N3)
}
\description{
The Alpha function is to generate valid significant level grid values
}
