% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{alpha_split}
\alias{alpha_split}
\title{This function is to fit a smooth model given alpha and corresponding power values from Monte Carlo sampling, and in 3-dim set, we suggest thin plate splines}
\usage{
alpha_split(
  r = c(1, 0.5, 0.3),
  sd = 1/base::sqrt(20),
  N1 = 20480,
  N2 = 10240,
  N3 = 2000,
  lower_bio_eff = 0.2,
  upper_bio_eff = 0.8,
  power,
  seed = NULL
)
}
\description{
This function is to fit a smooth model given alpha and corresponding power values from Monte Carlo sampling, and in 3-dim set, we suggest thin plate splines
}
