% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{bio_effect}
\alias{bio_effect}
\title{This function if to calculate the drug effect in each sub-population given the proportion}
\usage{
bio_effect(r, lower_bio_eff, upper_bio_eff)
}
\description{
This function if to calculate the drug effect in each sub-population given the proportion
}
