% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_function.R
\name{constraint}
\alias{constraint}
\title{The constraint function of significant level, typically is 0.025 for 2-side test}
\usage{
constraint(alpha_x, alpha_1n, r, sig.lv)
}
\description{
The constraint function of significant level, typically is 0.025 for 2-side test
}
