% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venv.R
\name{create_venv}
\alias{create_venv}
\title{A utility function to create and setup reticulate virtual environment and source the python function into the R environment.}
\usage{
create_venv(reticulate_venv = "designctpb_numba", py_path = NULL)
}
\description{
A utility function to create and setup reticulate virtual environment and source the python function into the R environment.
}
\examples{
create_venv()

}
