% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_ctpb.R
\name{design_ctpb}
\alias{design_ctpb}
\title{This function is to obtain fig.1 and get the optimal cutoff for the clinical design}
\usage{
design_ctpb(
  m = 24,
  n_dim = 3,
  N1 = 20480,
  N2 = 10240,
  N3 = 2000,
  E = NULL,
  SIGMA = NULL,
  sd_full = 1/base::sqrt(20),
  DELTA = NULL,
  delta_linear_bd = c(0.2, 0.8),
  seed = NULL
)
}
\description{
This function is to obtain fig.1 and get the optimal cutoff for the clinical design
}
