% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clinical.R
\name{plot_clinical}
\alias{plot_clinical}
\title{This function is to obtain fig.1 and get the optimal cutoff for the clinical design}
\usage{
plot_clinical(
  m = 19,
  n_dim = 3,
  sd = 1/base::sqrt(20),
  N1 = 20480,
  N2 = 10240,
  N3 = 2000,
  lower_bio_eff = 0.2,
  upper_bio_eff = 0.3
)
}
\arguments{
\item{m}{the number of points for r setting in each dimension}

\item{n_dim}{the number of dimension}

\item{sd}{the prior information of standard deviation for formula 11}

\item{N1}{Monte Carlo setting for calculating the density function f(Delta)}

\item{N2}{Monte Carlo setting for calculating the multivariate normal distribution in formula 1}

\item{N3}{the number of grid point for the sig.lv}

\item{upper_bio_eff}{the limit bio-marker effect when r approaches 0}

\item{low_bio_eff}{the bio-marker effect for the full population}

\item{power}{python function to estimate power values}
}
\description{
This function is to obtain fig.1 and get the optimal cutoff for the clinical design
}
\examples{
res <- plot_clinical()
res$plot_power # to see 3-d plot for the optimal power versus r2 and r3
res$plot_alpha # to see 3-d plot for the optimal alpha versus r2 and r3
res$opt_r_split #  to see the optimal cutoff of the sub-population, and here suggesting not cutoff at the 2-nd sub-population
res$opt_power 
res$opt_alpha_split
plot_clinical(m = 12, n_dim = 4, sd = 1/sqrt(40), N1 = 20480, N2 = 10240, N3 = 4000, lower_bio_eff=0.2, opper_bio_eff=0.8,power)# users could choose proper setting of m and N3, but we suggest not change N1 or N2, because the change should be applied in power4R.py too.

}
