% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{power_estimate_point}
\alias{power_estimate_point}
\title{This function is to estimate the power values given fixed r and the valid sig.lv matrix}
\usage{
power_estimate_point(r, sd, N1, N2, N3, lower_bio_eff, upper_bio_eff, power)
}
\arguments{
\item{r}{the proportion for each sub-population, r_1 is 1}

\item{sd}{the prior information of standard deviation for formula 11}

\item{N1}{Monte Carlo setting for calculating the density function f(Delta)}

\item{N2}{Monte Carlo setting for calculating the multivariate normal distribution in formula 1}

\item{N3}{the number of grid point for the sig.lv}

\item{upper_bio_eff}{the limit bio-marker effect when r approaches 0}

\item{power}{the python function to estimate power values}

\item{low_bio_eff}{the bio-marker effect for the full population}
}
\description{
This function is to estimate the power values given fixed r and the valid sig.lv matrix
}
\examples{
reticulate::source_python("power4R.py")
power = Power_sampling
power_estimate_point(r,sd,N1,N2,N3,lower_bio_eff, upper_bio_eff,power)

}
