% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{power_estimator}
\alias{power_estimator}
\title{Estimate the power for N3 alpha, given fixed r}
\usage{
power_estimator(
  r,
  N1,
  N2,
  N3,
  E = NULL,
  sig = NULL,
  sd_full,
  delta = NULL,
  delta_linear_bd,
  Power = NULL,
  seed = NULL
)
}
\description{
Estimate the power for N3 alpha, given fixed r
}
