% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{proportion}
\alias{proportion}
\title{Grid setting of proportions for each sub-population}
\usage{
proportion(m, n_dim)
}
\arguments{
\item{m}{integer, the number of grid points in each dimension, and we suggest m around 20 for 3 dimension}

\item{n_dim}{integer for the dimension, which is equal to the number of sub-population plus 1}
}
\value{
matrix of setting the proportion of the population by given specific dimension and density in each dimension
}
\description{
This function is to decide the r setting given specific density in each dimension
}
