% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_ini.R
\name{py_ini}
\alias{py_ini}
\title{Initiate the reticulate environment}
\usage{
py_ini()
}
\value{
Error in CudaSupport means there is no cuda device or cuda driver lib cannot be found; otherwise a message of successfully set up the environment will be shown
}
\description{
A utility function to initiate the python environment , check CUDA device and source the python function into the R environment and set it as global function.
}
\examples{
\dontrun{
py_ini()
}
}
