% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venv.R
\name{py_initial}
\alias{py_initial}
\title{A utility function to source the python function into the R environment.}
\usage{
py_initial(reticulate_venv = "designctpb_numba")
}
\description{
A utility function to source the python function into the R environment.
}
\examples{
py_initial()

}
