% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{r_setting}
\alias{r_setting}
\title{This function is to decide the r setting given specific density in each dimension}
\usage{
r_setting(m, n_dim)
}
\description{
This function is to decide the r setting given specific density in each dimension
}
