\name{res_strong}
\alias{res_strong}
\docType{data}
\title{
The clinical trial design for strong biomarker effect
}
\description{
The dataset contains the 3-D rotable figure for the simulated nested two-subset design(strong biomarker effect), from which the optimal choice of each population's proportion and the optimal alpha split can be abtianed. 
}
\usage{data("res_strong")}
\format{
  The format is:
List of 5
 $ plot_power     : plotly object
 $ plot_alpha     : plotly object
 $ opt_r_split    : numeric vector
 $ opt_power      : numeric 
 $ opt_alpha_split: numeric vector 
}
\examples{
data(res_strong)
}
\keyword{datasets}
