% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smooth_power_alpha}
\alias{smooth_power_alpha}
\title{This function is to fit a smooth model given alpha and corresponding power values from Monte Carlo sampling, and in 3-dim set, we suggest thin plate splines}
\usage{
smooth_power_alpha(r, sd, N1, N2, N3, lower_bio_eff, upper_bio_eff, power)
}
\arguments{
\item{r}{the proportion for each sub-population, r_1 is 1}

\item{sd}{the prior information of standard deviation for formula 11}

\item{N1}{Monte Carlo setting for calculating the density function f(Delta)}

\item{N2}{Monte Carlo setting for calculating the multivariate normal distribution in formula 1}

\item{N3}{the number of grid point for the sig.lv}

\item{upper_bio_eff}{the limit bio-marker effect when r approaches 0}

\item{power}{python function to estimate power}

\item{low_bio_eff}{the bio-marker effect for the full population}
}
\description{
This function is to fit a smooth model given alpha and corresponding power values from Monte Carlo sampling, and in 3-dim set, we suggest thin plate splines
}
