% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{Alpha}
\alias{Alpha}
\title{The Alpha function is to generate valid significant level grid values}
\usage{
Alpha(r, N3)
}
\arguments{
\item{r}{the proportion for each sub-population, r_1 is 1}

\item{N3}{the number of grid point for the sig.lv}
}
\description{
The Alpha function is to generate valid significant level grid values
}
