% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_function.R
\name{alpha_kernel}
\alias{alpha_kernel}
\title{The function is calculating the significant level for the n-th sub-population subject to the constraint function}
\usage{
alpha_kernel(alpha_tol, r, sig.lv)
}
\description{
The function is calculating the significant level for the n-th sub-population subject to the constraint function
}
