% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{alpha_split}
\alias{alpha_split}
\title{This function is to fit a smooth model given alpha and corresponding power values from Monte Carlo sampling, and in 3-dim set, we suggest thin plate splines}
\usage{
alpha_split(
  r,
  sd,
  N1,
  N2,
  N3,
  lower_bio_eff,
  upper_bio_eff,
  power,
  seed = NULL
)
}
\description{
This function is to fit a smooth model given alpha and corresponding power values from Monte Carlo sampling, and in 3-dim set, we suggest thin plate splines
}
