% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_function.R
\name{constraint}
\alias{constraint}
\title{Constraint function}
\usage{
constraint(alpha_x, alpha_1n, r, sig.lv)
}
\arguments{
\item{alpha_x}{sig.lv for the n-th(smallest) sub-population}

\item{alpha_1n}{vector of sig.lv for the full to (n-1)-th(except for the smallest) sub-population}

\item{r}{the proportion for each sub-population, r_1 is 1, r_i>r_{i+1}}

\item{sig.lv}{significant level for hypothesis testing, usually 0.025 for 2-side test}
}
\value{
the result is 0, refer to the Formula(2) of preprint paper arXiv:2005.10494
}
\description{
The constraint function of significant level for the nested-population to test the problem, typically 0.025 for two-side test and 0.05 for one-side test.
}
