% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venv.R
\name{create_venv}
\alias{create_venv}
\title{Setup reticulate environment}
\usage{
create_venv(reticulate_venv, py_path = NULL)
}
\arguments{
\item{reticulate_venv}{character, the name of the reticulate virtual environment to setup.}

\item{py_path}{character, python path, the default value is NULL which the default python path will be applied}
}
\description{
A utility function to create and setup reticulate virtual environment and source the python function into the R environment.
}
\details{
set up the python environment, and check whether the required python modules installed or not, if not install it.
}
\examples{
create_venv()

}
