% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{Optim_Res}
\alias{Optim_Res}
\title{This function is to obtain the optimal results given grid points of r setting}
\usage{
Optim_Res(
  m,
  r_set,
  n_dim,
  N1,
  N2,
  N3,
  E,
  SIGMA,
  sd_full,
  DELTA,
  delta_linear_bd,
  Power,
  seed
)
}
\description{
This function is to obtain the optimal results given grid points of r setting
}
