% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{optim_res}
\alias{optim_res}
\title{This function is to obtain the optimal results given grid points of r setting}
\usage{
optim_res(
  m,
  n_dim,
  sd,
  N1,
  N2,
  N3,
  lower_bio_eff,
  upper_bio_eff,
  power,
  seed = NULL
)
}
\description{
This function is to obtain the optimal results given grid points of r setting
}
