% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_ini.R
\name{py_ini}
\alias{py_ini}
\title{Initiate the reticulate environment}
\usage{
py_ini()
}
\value{
the physical address of R function Power_sampling, converted from python file "python4R.py"
}
\description{
A utility function to initiate the python environment and source the python function into the R environment and set it as global function.
}
\examples{
\dontrun{
py_ini()
}
}
