% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venv.R
\name{py_initial}
\alias{py_initial}
\title{Initiate the reticulate environment}
\usage{
py_initial(reticulate_venv)
}
\arguments{
\item{reticulate_venv}{character, the name of the reticulate virtual environment has been setup}
}
\description{
A utility function to initiate the python environment created by create_venv(), source the python function into the R environment and set it as global function.
}
\examples{
py_initial()

}
